add fe ut
