+++ fe/fe-core/src/test/java/com/starrocks/catalog/JDBCTableTest.java
@@ -15,6 +15,7 @@
 
 package com.starrocks.catalog;
 
+import com.google.common.collect.ImmutableMap;
 import com.google.common.collect.Lists;
 import com.google.common.collect.Maps;
 import com.starrocks.common.DdlException;
@@ -29,6 +30,7 @@ import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
 
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
 
@@ -217,5 +219,64 @@ public class JDBCTableTest {
         properties.remove("table");
         new JDBCTable(1000, "jdbc_table", columns, properties);
         Assert.fail("No exception throws.");
+    }
+
+    @Test
+    public void testJDBCDriverName() {
+        try {
+            Map<String, String> properties = ImmutableMap.of(
+                    "driver_class", "org.postgresql.Driver",
+                    "checksum", "bef0b2e1c6edcd8647c24bed31e1a4ac",
+                    "driver_url",
+                    "http://x.com/postgresql-42.3.3.jar",
+                    "type", "jdbc",
+                    "user", "postgres",
+                    "password", "postgres",
+                    "jdbc_uri", "jdbc:postgresql://172.26.194.237:5432/db_pg_select"
+            );
+            List<Column> schema = new ArrayList<>();
+            schema.add(new Column("id", Type.INT));
+            JDBCTable jdbcTable = new JDBCTable(10, "tbl", schema, "db", "jdbc_catalog", properties);
+            TTableDescriptor tableDescriptor = jdbcTable.toThrift(null);
+            TJDBCTable table = tableDescriptor.getJdbcTable();
+            Assert.assertEquals(table.getJdbc_driver_name(), "jdbc_postgresql___172_26_194_237_5432_db_pg_select");
+            Assert.assertEquals(table.getJdbc_driver_url(), "http://x.com/postgresql-42.3.3.jar");
+            Assert.assertEquals(table.getJdbc_driver_checksum(), "bef0b2e1c6edcd8647c24bed31e1a4ac");
+            Assert.assertEquals(table.getJdbc_driver_class(), "org.postgresql.Driver");
+        } catch (Exception e) {
+            System.out.println(e.getMessage());
+            Assert.fail();
+        }
+    }
+
+    @Test
+    public void testJDBCDriverNameLong() {
+        try {
+            Map<String, String> properties = ImmutableMap.of(
+                    "driver_class", "org.postgresql.Driver",
+                    "checksum", "bef0b2e1c6edcd8647c24bed31e1a4ac",
+                    "driver_url",
+                    "http://x.com/postgresql-42.3.3.jar",
+                    "type", "jdbc",
+                    "user", "postgres",
+                    "password", "postgres",
+                    "jdbc_uri",
+                    "jdbc:postgresql" +
+                            "://aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa.com/db_pg_select"
+            );
+            List<Column> schema = new ArrayList<>();
+            schema.add(new Column("id", Type.INT));
+            JDBCTable jdbcTable = new JDBCTable(10, "tbl", schema, "db", "jdbc_catalog", properties);
+            TTableDescriptor tableDescriptor = jdbcTable.toThrift(null);
+            TJDBCTable table = tableDescriptor.getJdbcTable();
+            Assert.assertEquals(table.getJdbc_driver_name(),
+                    "jdbc_90377bb27298feecdca1ef8b2e9c2e00f0b012eabb9ad43437542d2e29ef52fc");
+            Assert.assertEquals(table.getJdbc_driver_url(), "http://x.com/postgresql-42.3.3.jar");
+            Assert.assertEquals(table.getJdbc_driver_checksum(), "bef0b2e1c6edcd8647c24bed31e1a4ac");
+            Assert.assertEquals(table.getJdbc_driver_class(), "org.postgresql.Driver");
+        } catch (Exception e) {
+            System.out.println(e.getMessage());
+            Assert.fail();
+        }
     }
 }
