+++ fe/fe-core/src/test/java/com/starrocks/connector/jdbc/JDBCTableTest.java
@@ -14,18 +14,9 @@
 
 package com.starrocks.connector.jdbc;
 
-import com.google.common.collect.ImmutableMap;
-import com.starrocks.catalog.Column;
-import com.starrocks.catalog.JDBCTable;
-import com.starrocks.catalog.Type;
-import com.starrocks.thrift.TJDBCTable;
-import com.starrocks.thrift.TTableDescriptor;
 import org.junit.Assert;
 import org.junit.Test;
 
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Map;
 import java.util.Objects;
 
 public class JDBCTableTest {
@@ -52,34 +43,6 @@ public class JDBCTableTest {
             Assert.assertTrue(partition.hashCode() == Objects.hash("20230810", 1000L));
             Assert.assertTrue(partition.toString().contains("20230810"));
             Assert.assertTrue(partition.toJson().toString().contains("20230810"));
-        } catch (Exception e) {
-            System.out.println(e.getMessage());
-            Assert.fail();
-        }
-    }
-
-    @Test
-    public void testJDBCDriverName() {
-        try {
-            Map<String, String> properties = ImmutableMap.of(
-                    "driver_class", "org.postgresql.Driver",
-                    "checksum", "bef0b2e1c6edcd8647c24bed31e1a4ac",
-                    "driver_url",
-                    "http://x.com/postgresql-42.3.3.jar",
-                    "type", "jdbc",
-                    "user", "postgres",
-                    "password", "postgres",
-                    "jdbc_uri", "jdbc:postgresql://172.26.194.237:5432/db_pg_select"
-            );
-            List<Column> schema = new ArrayList<>();
-            schema.add(new Column("id", Type.INT));
-            JDBCTable jdbcTable = new JDBCTable(10, "tbl", schema, "db", "jdbc_catalog", properties);
-            TTableDescriptor tableDescriptor = jdbcTable.toThrift(null);
-            TJDBCTable table = tableDescriptor.getJdbcTable();
-            Assert.assertEquals(table.getJdbc_driver_name(), "jdbc_postgresql_172.26.194.237_5432_db_pg_select");
-            Assert.assertEquals(table.getJdbc_driver_url(), "http://x.com/postgresql-42.3.3.jar");
-            Assert.assertEquals(table.getJdbc_driver_checksum(), "bef0b2e1c6edcd8647c24bed31e1a4ac");
-            Assert.assertEquals(table.getJdbc_driver_class(), "org.postgresql.Driver");
         } catch (Exception e) {
             System.out.println(e.getMessage());
             Assert.fail();
